Notifiers for Logger, separate MailNotifier class
