+++ src/Tracy/INotifier.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * This file is part of the Tracy (https://tracy.nette.org)
+ * Copyright (c) 2004 David Grudl (https://davidgrudl.com)
+ */
+
+namespace Tracy;
+
+
+/**
+ * Notifier interface.
+ */
+interface INotifier
+{
+
+	public function notify($message, $priority);
+
+}
