+++ src/Tracy/Logger.php
@@ -25,19 +25,25 @@ class Logger implements ILogger
 	/** @var mixed interval for sending email is 2 days */
 	public $emailSnooze = '2 days';
 
-	/** @var callable handler for sending emails */
-	public $mailer;
-
 	/** @var BlueScreen */
 	private $blueScreen;
+
+	/** @var MessageFormat */
+	private $messageFormat;
+
+	/** @var array of INotifier instances */
+	private $notifiers;
 
 
 	public function __construct($directory, $email = NULL, BlueScreen $blueScreen = NULL)
 	{
 		$this->directory = $directory;
-		$this->email = $email;
 		$this->blueScreen = $blueScreen;
-		$this->mailer = [$this, 'defaultMailer'];
+		$this->messageFormat = new MessageFormat();
+
+		$mailNotifier = new MailNotifier($email, $directory);
+
+		$this->notifiers = [$mailNotifier];
 	}
 
 
@@ -58,7 +64,7 @@ class Logger implements ILogger
 		$exceptionFile = $message instanceof \Exception || $message instanceof \Throwable
 			? $this->getExceptionFile($message)
 			: NULL;
-		$line = $this->formatLogLine($message, $exceptionFile);
+		$line = $this->messageFormat->formatLogLine($message, $exceptionFile);
 		$file = $this->directory . '/' . strtolower($priority ?: self::INFO) . '.log';
 
 		if (!@file_put_contents($file, $line . PHP_EOL, FILE_APPEND | LOCK_EX)) { // @ is escalated to exception
@@ -69,50 +75,11 @@ class Logger implements ILogger
 			$this->logException($message, $exceptionFile);
 		}
 
-		if (in_array($priority, [self::ERROR, self::EXCEPTION, self::CRITICAL], TRUE)) {
-			$this->sendEmail($message);
+		foreach ($this->notifiers as $notifier) {
+			$notifier->notify($message, $priority);
 		}
 
 		return $exceptionFile;
-	}
-
-
-	/**
-	 * @param  string|\Exception|\Throwable
-	 * @return string
-	 */
-	protected function formatMessage($message)
-	{
-		if ($message instanceof \Exception || $message instanceof \Throwable) {
-			while ($message) {
-				$tmp[] = ($message instanceof \ErrorException
-					? Helpers::errorTypeToString($message->getSeverity()) . ': ' . $message->getMessage()
-					: Helpers::getClass($message) . ': ' . $message->getMessage()
-				) . ' in ' . $message->getFile() . ':' . $message->getLine();
-				$message = $message->getPrevious();
-			}
-			$message = implode($tmp, "\ncaused by ");
-
-		} elseif (!is_string($message)) {
-			$message = Dumper::toText($message);
-		}
-
-		return trim($message);
-	}
-
-
-	/**
-	 * @param  string|\Exception|\Throwable
-	 * @return string
-	 */
-	protected function formatLogLine($message, $exceptionFile = NULL)
-	{
-		return implode(' ', [
-			@date('[Y-m-d H-i-s]'), // @ timezone may not be set
-			preg_replace('#\s*\r?\n\s*#', ' ', $this->formatMessage($message)),
-			' @  ' . Helpers::getSource(),
-			$exceptionFile ? ' @@  ' . basename($exceptionFile) : NULL,
-		]);
 	}
 
 
@@ -134,6 +101,16 @@ class Logger implements ILogger
 
 
 	/**
+	 * @param  \Tracy\INotifier $notifier
+	 * @return void
+	 */
+	public function addNotifier(INotifier $notifier)
+	{
+		$this->notifiers[] = $notifier;
+	}
+
+
+	/**
 	 * Logs exception to the file if file doesn't exist.
 	 * @param  \Exception|\Throwable
 	 * @return string logged error filename
@@ -144,54 +121,6 @@ class Logger implements ILogger
 		$bs = $this->blueScreen ?: new BlueScreen;
 		$bs->renderToFile($exception, $file);
 		return $file;
-	}
-
-
-	/**
-	 * @param  string|\Exception|\Throwable
-	 * @return void
-	 */
-	protected function sendEmail($message)
-	{
-		$snooze = is_numeric($this->emailSnooze)
-			? $this->emailSnooze
-			: @strtotime($this->emailSnooze) - time(); // @ timezone may not be set
-
-		if ($this->email && $this->mailer
-			&& @filemtime($this->directory . '/email-sent') + $snooze < time() // @ file may not exist
-			&& @file_put_contents($this->directory . '/email-sent', 'sent') // @ file may not be writable
-		) {
-			call_user_func($this->mailer, $message, implode(', ', (array) $this->email));
-		}
-	}
-
-
-	/**
-	 * Default mailer.
-	 * @param  string|\Exception|\Throwable
-	 * @param  string
-	 * @return void
-	 * @internal
-	 */
-	public function defaultMailer($message, $email)
-	{
-		$host = preg_replace('#[^\w.-]+#', '', isset($_SERVER['HTTP_HOST']) ? $_SERVER['HTTP_HOST'] : php_uname('n'));
-		$parts = str_replace(
-			["\r\n", "\n"],
-			["\n", PHP_EOL],
-			[
-				'headers' => implode("\n", [
-					'From: ' . ($this->fromEmail ?: "noreply@$host"),
-					'X-Mailer: Tracy',
-					'Content-Type: text/plain; charset=UTF-8',
-					'Content-Transfer-Encoding: 8bit',
-				]) . "\n",
-				'subject' => "PHP: An error occurred on the server $host",
-				'body' => $this->formatMessage($message) . "\n\nsource: " . Helpers::getSource(),
-			]
-		);
-
-		mail($email, $parts['subject'], $parts['body'], $parts['headers']);
 	}
 
 }
